v0.1.4 Stable
FIXES [TO READ ] :
—————————————
The problem that caused the ‘OneEntitySelectMenu’ to work only when an
entity had values saved was due to the lack of setting the process
attribut on  CommandButtons ‘Nouveau’ & ‘Modifier’ so that they process
only the concerned OneEntitySelectMenu and not the full mainForm.
Actually when the full mainForm was processed at click of the one these
two buttons, and the values were not set correctly ( missing required
values …) error messages were generated but not shown and this caused
the rest of code that would show the dialog with the correct values …
etc to not been show.
Actually, it would be optimal to set process attribute on command
‘Button of  dialog forms so that process only that dialog and not the
full mainForm. If the process target is not set, the save button might
not work if there is any other dialogForm that is non-filled, because
it would be process too within any other part of the mainForm.

Many Other Enhancements :
—————————————
- Not more need to specify the name of the handled bean on dialog
forms, this value is actually declared one time at _config.xhtml file
- Other minor design & code fixes