v0.1.5 Stable
Very Important FIXE
——————————
1. Added attribute process to commandButton ‘Enregistrer’
process=":dialog#{entityNameCapitalized}. This apply the submit only to
the dialog concerned part. So that if many other subForms are added
they would not block the current submit request.
This is an optimisation also, the submit theoretically would run
faster, because on the small concerned part is being submitted.

2. Also, made the id of dialog unique by adding the the variable’s
‘entityNameCapitalized’ value to it :
id="dialogForm#{entityNameCapitalized}">