v0.2.1 Stable
ADDS / FIXES
——————————
1. Fixed ContextMenuItem ‘Modifier’ redirection so that it always
passes the parameter within the url, it’s no more based on setting the
UI param on the contextMenu Item. It’s now using an action listener to
execute the redirection and the parameter ID set correctly, through the
java method.
GenricCRUDController::_actions.get(‘redirectToFormView’)
2. Factorized the properties ‘_moduleName’ and ‘_actions’ that were
common between the subclasses of GenericController to GenericController
so they are written only one time.
3. Some UI Enhancements : Dev Tools Box… etc