v0.3.0 Stable
FIXE
———
FIXED the update log of OneEntitySelectMenu at prePostQuickEdit event.
No more need to specify the update code, the code now is cleaner, the
only part need is how to set and get the related entity value.

BUG TO CORRECT
—————————

There is still a bug when multiple OneEntitySelectMenu is called. Only
the last one works